+++ python_training/Midas/midas/core/engine.py
@@ -1,5 +1,5 @@
 from sqlalchemy import create_engine
 
-from python_training.midas.config import OPERATIONAL_DB
+from midas.conf.config import OPERATIONAL_DB
 
 engine = create_engine(OPERATIONAL_DB, echo=True)
