+++ python_training/Midas/midas/core/event_member_enrollment.py
@@ -1,6 +1,6 @@
 from sqlalchemy import Table, Column, Integer, ForeignKey
 
-from python_training.midas.base import Base
+from midas.core.base import Base
 
 event_enrollment_association_table = Table('event_member_association', Base.metadata,
                                            Column('event_id', Integer,
