+++ python_training/Midas/midas/core/organization.py
@@ -1,8 +1,8 @@
-from python_training.base import Base
+from midas.core.base import Base
 from sqlalchemy import Column, Integer, String
 from sqlalchemy.orm import relationship
 
-from python_training.midas.basic_table import BasicTable
+from midas.core.basic_table import BasicTable
 
 
 class Organization(Base, BasicTable):
